Move remaining repos/services
This moves the leftover services and repos into either a feature
app location of a general library.

It also removes services that no longer seem to have a use.
