+++ src/apps/companies/controllers/ltd.controller.js
@@ -2,7 +2,7 @@ const companyService = require('../services/data.service')
 const companyFormService = require('../services/form.service')
 const companyFormattingService = require('../services/formatting.service')
 const companyRepository = require('../repository')
-const metadataRepository = require('../../../repos/metadata.repo')
+const metadataRepository = require('../../../lib/metadata')
 const { companyDetailsLabels, chDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels')
 const { containsFormData, isBlank } = require('../../../lib/controller-utils')
 const companyWithCHKeys = ['alias', 'trading_address', 'uk_region', 'headquarter_type', 'sector', 'website', 'description', 'employee_range', 'turnover_range']
