+++ src/apps/contacts/services/data.service.js
@@ -1,8 +1,8 @@
 /* eslint camelcase: 0 */
 const logger = require('../../../../config/logger')
-const adviserRepository = require('../../../repos/adviser.repo')
+const adviserRepository = require('../../adviser/repository')
 const interactionRepository = require('../../interactions/interactions.repo')
-const metadataRepository = require('../../../repos/metadata.repo')
+const metadataRepository = require('../../../lib/metadata')
 const serviceDeliveryRepository = require('../../service-deliveries/service-deliveries.repo')
 
 /**
