+++ src/apps/interactions/controllers/edit.controller.js
@@ -2,8 +2,8 @@
 const interactionLabels = require('../labels')
 const contactsRepository = require('../../contacts/contacts.repo')
 const companyRepository = require('../../companies/repository')
-const metadataRepository = require('../../../repos/metadata.repo')
-const adviserRepository = require('../../../repos/adviser.repo')
+const metadataRepository = require('../../../lib/metadata')
+const adviserRepository = require('../../adviser/repository')
 const interactionDataService = require('../services/data.service')
 const interactionFormService = require('../services/form.service')
 const { containsFormData } = require('../../../lib/controller-utils')
