+++ src/server.js
@@ -15,7 +15,7 @@ const nunjucks = require('../config/nunjucks')
 const datahubFlash = require('./middleware/flash')
 const headers = require('./middleware/headers')
 const locals = require('./middleware/locals')
-const metadata = require('./repos/metadata.repo')
+const metadata = require('./lib/metadata')
 const user = require('./middleware/user')
 const auth = require('./middleware/auth')
 const csrfToken = require('./middleware/csrf-token')
