--- src/services/remote-company.service.js
@@ -1,10 +1,0 @@
-const axios = require('axios')
-
-function getCompanyForSource (token = null, id, source) {
-  return axios.get(`/api/company/${source}/${id}`)
-    .then(response => response.data)
-}
-
-module.exports = {
-  getCompanyForSource,
-}
