--- src/services/remote-contact.service.js
@@ -1,10 +1,0 @@
-const axios = require('axios')
-
-function getInflatedContact (token = null, id) {
-  return axios.get(`/api/contact/${id}`)
-    .then(response => response.data)
-}
-
-module.exports = {
-  getInflatedContact,
-}
