+++ test/unit/apps/companies/controllers/ukother.controller.test.js
@@ -72,7 +72,7 @@ describe('Company controller, uk other', function () {
         getCHCompany: getCHCompanyStub,
         getDitCompany: getDitCompanyStub,
       },
-      '../../../repos/metadata.repo': metadataRepositoryStub,
+      '../../../lib/metadata': metadataRepositoryStub,
     })
   })
 
@@ -501,7 +501,7 @@ describe('Company controller, uk other', function () {
           getCHCompany: getCHCompanyStub,
           getDitCompany: getDitCompanyStub,
         },
-        '../../../../repos/metadata.repo': metadataRepositoryStub,
+        '../../../lib/metadata': metadataRepositoryStub,
       })
 
       const body = {
